Enable upload of continuous package by fixing a typo
* This variable is used to set the FOLDERID parameter of the
"midas_api_item_upload" function. As a visible result, the error reported
when running the dashboard script mentioned that the "midas*_response.txt"
has to be analyzed. This one contain a message stating that
"Parameter filename is not defined".

Long story short .. the module MIDASAPICore will have to be tested
for invalid parameter.