STYLE: Refactor platform conditionals
Change preprocessor logic to select platform specific code to
consistently use the Qt OS symbols, to treat all non-Windows platforms
as POSIX, and to explicitly fail (i.e. '#error') in case of an unknown
platform. This simplifies the preprocessor logic and produces a better
error in case of an unsupported platform (which previously would have
failed due to use of unknown variables).