ENH: Support additional path list envvars
Add support for an additional list of environment variables that should
be treated as path lists. Add ability to specify these as lists in the
manner of [Paths] and [LibraryPaths] (which are merged with any values
set in the "traditional" manner for backwards compatibility). Refactor
buildEnvironment to treat PATH and the platform-appropriate library path
in the same manner as these (which greatly simplifies the code).

Besides the added convenience in the configuration file, the generated
wrapper script now also prepends to these variables as well as the usual
PATH and library path variables.