BUG: Fix ctkAppLauncherConfigure dependencies
Change rule to generate the launcher configuration to actually depend on
the input files (i.e. the scripts that contain the configuration files).
This fixes failure to regenerate the configuration when changes are made
at the CMake level.

Also, add some comments and fix the script dependency for generating the
configurations, which should be the script we invoke to do so, not the
script which contains the macro to create the target (which is in no way
useful as a dependency; at least not with the new, correct
dependencies).

Note that this is a little (but harmlessly) over-zealous right now, as
it forces regenerating the configuration any time the project is
reconfigured (as the input settings files are currently being written
unconditionally). This is addressed by f64df2a0f515 in a separate
branch, where writing the settings files is changed to only update the
files if the contents would change.