BUG: "Fix" creation of launcher executable
Remove rule in ctkAppLauncher-configure.cmake to create the launcher
executable in favor of a build rule set up by ctkAppLauncher.cmake. This
reduces dependency twining, and the dependencies should now be correct
to update the executable if the source instance changes or if it has
been deleted, without the launcher configuration also needing to be
regenerated.