ENH: Copy CMake helper scripts to build directory
Add rules to copy the CMake helper scripts to the build directory. This
should allow users (i.e. projects creating their own launchers) to use
the launcher from the build directory as well as an install directory.
(In particular, we want to write a unit test for ctkAppLauncherConfigure
which will do so. This requires that the CMake helper scripts and the
built launcher executable be in the same tree.)