BUG: Fix problem when no extra paths set
Launch configurations will likely store an empty list of additional path
variables as an empty value. However, this is converted to a list
containing an empty string, which makes ctk::readArrayValues unhappy
when asked to read entries for a list with an empty name. Therefore,
only consider the name if it is non-empty.

See also https://bugreports.qt-project.org/browse/QTBUG-19823.