ENH: Add unit tests for additional paths feature
Modify the new configure unit test from the other branch to also test
the new additional paths feature. Modify the TestEnvironment unit tests
(and the test application) to also test the same.

Note that TestEnvironment-NotFound actually had another bug that is now
fixed; the third expected value name was wrong, and was thus not
expanding, and as a result we never actually tested the value. The new
regex has an additional trailing newline that will prevent that sort of
thing from happening (as long as the value doesn't start with a newline,
anyway).