BUG: Fix App4ConfigureLauncherTest-ConfigureStep
Tweak ctkAppLauncher.cmake to accept an optional argument with build
configurations to use as hints when looking for the launcher executable
(which is needed to find it in the build dirextory of a multi-config
build). Adjust test invocation to pass along current configuration to
the test scripts, and aforementioned test to pass along the
configuration when configuring the test project, for use in the new
search path list.

This fixes App4ConfigureLauncherTest-ConfigureStep failing with
multi-config generators.