BUG: Fix TestLauncherVerbose test
Fix regex used in TestLauncherVerbose looking for "LD_LIBRARY_PATH",
which is not present on OS/X or Windows, to look for the appropriate
variable. (We're a little lax; we'll take either "LD_LIBRARY_PATH" or
"DYLD_LIBRARY_PATH" on not-Windows. On Windows, we'll look for "PATH".
This should be close enough given we're really looking for logging in
general and not so much specific logging messages except as that's the
easiest way to check for logging output.)