BUG: Correctly fix tests on Windows
Quote strings where needed to prevent expected/actual output comparison
from incorrectly removing ';'s. It turns out this was the actual problem
with these tests, and not line endings as previously believed (see
8011e58756b6). The old work-around (now removed) happened to "fix" the
problem by also stripping ';'s from the expected string, which allowed
the compare to succeed and the test to pass.