Add testing Dockerfile.
The tests run in a Docker image that has a graphical context
(thewtex/opengl). Set the test command for that container with the
environmental variable GRAPHICAL_APP. CMake is added with the same version and
Tthe build container so ctest can be executed.

To run, a version of the applauncher build container should be created, and
the volumes mounted with '--volumes-from'.

  cd AppLauncher
  docker build -t applauncher .
  docker create --name applauncher-container applauncher /bin/true
  cd Testing/Docker
  docker build -t applauncher-test .
  docker run --name applauncher-test-container \
    --volumes-from applauncher-container \
    applauncher-test
  docker cp applauncher-test-container:/var/log/supervisor/graphical-app-launcher.log - | tar xO