Add CirceCI configuration.
The 'dependencies' section pulls DockerHub archives of the Docker images to
reduce build time and avoid loss of dependencies like the Qt layer.

The 'test' section builds the Dockerfile's in the repository used to both
package the application on run the test suit. Next, the current pull request
is build and packaged. Next, the generated package is moved to the CircleCI
artifacts directory for retrieval after the build. Then, 'ctest' is executed
on the build tree in the test image, which has graphical support.

Finally, in the 'deployment' section, the built images are archived on
DockerHub.