CTKAppLauncherLibConfig: Do not force use of Qt4 imported targets
Since on MacOSX not all modules are available as "framework" (e.g libQtUiTools.a),
the assumption made in "FindQt4.cmake" is not always valid and systematically
setting QT_USE_IMPORTED_TARGETS to 1 will cause build error.

Indeed, it turns out that FindQt4.cmake expects all libraries to be
available as framework if at least "QtCore.framework" is available but this
is not true for QtUiTools.