BUG: fix crash calculating environment keys
Fixes #116

Each call to `ctkAppLauncherEnvironment::excludeReservedVariableNames` returns a new `QStringList` so making a QSet from the `begin()` of one and the `end()` of the other led to the crash in the `QHash`.

Solution is to make temporary variables for the `QStringList`s so that the set is made from matching iterators variables.

Also rearranged the code a bit for readability.