ENH: Allow user settings file to be in <APPLAUNCHER_DIR> (#114)
ENH: Allow user settings file to be within APPLAUNCHER_DIR

Before this commit, user settings file was always stored in the user profile directory, which made it impossible
to create completely self-contained, portable applications (which store all settings within <APPLAUNCHER_DIR> and subfolders).

This is fixed by two changes:

1. The launcher looks for UserAdditionalSettings file at <APPLAUNCHER_DIR>/<organizationName|organizationDomain>/<ApplicationName>(-<revision>).ini
first and if that file is found then that is used. If settings is not found there then the user's profile directory is used the same way as before.

2. Relative paths can be specified (they are resolved using <APPLAUNCHER_DIR> directory as the base).

Lastly, this commit adds option to msvc-static-configure.cmake to use new CMake and VS2019 (with VS2017 toolset)

The new option configures 'Visual Studio 16 2019' with VS2017 toolset and 32-bit build configuration.

There is no change for the existing 'Ninja' and 'Visual Studio 15 2017' options. Eventually, these options can be upgraded to work with current CMake version.

Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>