tests: Fix AppLauncher-TestSaveLauncherEnvironment by supporting optional launcher copy
This patch resolves a race condition and file locking issue affecting the
`AppLauncher-TestSaveLauncherEnvironment` test on Windows by introducing
the option `AppLauncherTestPrerequisites_COPY_LAUNCHER`.

It allows tests to disable copying of the launcher binary when including
`AppLauncherTestPrerequisites.cmake`, avoiding conflicts with in-use
executables on Windows.

It addresses errors like the following:

  file COPY_FILE failed to copy
    CTKAppLauncher.exe -> CTKAppLauncher_<hash>.exe
  because: Permission denied